Use a tempfile when importing city dump
The PostgreSQL docker container has provisions for a one-shot
database initialization. However, if we call the db bootstrap from
the postgres init, it will fail with errors trying to create the
city.dump file.

This commit adds a tempfile which is used to create and load
city.dump.

Normally, I would use a context for this, but when making changes
to OPC, I make the minimum change necessary.